backend/requirements.txt
fastapi
uvicorn
pydantic
prisma
google-generativeai
python-dotenv
asyncio
